automation/requirements.txt
pyyaml
toml
markdown
requests
python-dotenv
beautifulsoup4
